@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.1 <4.0.0"
 
 dependencies:
   cupertino_icons: ^1.0.5
