@@ -1,14 +1,14 @@
 name: google_places_flutter
 description: Custom Google places autocomplete widget for Android and iOS both. Use https://pub.dev/packages/google_places_flutter
-version: 2.1.0
+version: 2.2.0
 author: Shruti Mahajan <shruti.sm87@gmail.com>
 homepage: https://github.com/Shrutimahajan/Google-AutoComplete-TextField-Flutter
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.1 <4.0.0'
 
 dependencies:
-  dio: ^4.0.0  
+  dio: ^5.0.3  
   flutter:
     sdk: flutter
   rxdart: ^0.27.2
