@@ -8,9 +8,8 @@ import 'package:google_places_flutter/model/prediction.dart';
 void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -27,9 +26,9 @@ class MyHomePage extends StatefulWidget {
   final String title;
 
   const MyHomePage({
-    Key? key,
+    super.key,
     required this.title,
-  }) : super(key: key);
+  });
 
   @override
   State<MyHomePage> createState() => _MyHomePageState();
@@ -42,28 +41,10 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title),
       ),
       body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
         child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
           children: <Widget>[
             const SizedBox(height: 20),
             placesAutoCompleteTextField(),
@@ -83,7 +64,6 @@ class _MyHomePageState extends State<MyHomePage> {
               const InputDecoration(hintText: 'Search your location'),
           debounceTime: 800,
           countries: const <String>['in', 'fr'],
-          isLatLngRequired: true,
           getPlaceDetailWithLatLng: (Prediction prediction) {
             print('placeDetails${prediction.lng}');
           },
@@ -92,9 +72,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
             controller.selection = TextSelection.fromPosition(
                 TextPosition(offset: prediction.description?.length ?? 0));
-          }
-          // default 600 ms ,
-          ),
+          }),
     );
   }
 }
